:building_construction: Move storage key code from add to file
