+++ lamindb/dev/db/_add.py
@@ -1,15 +1,12 @@
 from functools import partial
-from pathlib import Path, PurePath
 from typing import List, Optional, Tuple, Union, overload  # noqa
 
 import sqlmodel as sqm
 from lamin_logger import logger
 from lndb import settings as setup_settings
-from lndb_storage import UPath, delete_storage, store_object, write_adata_zarr
+from lndb_storage import delete_storage, store_object, write_adata_zarr
 from lndb_storage._file import print_hook
-from lnschema_core import File, Folder
-from pydantic.fields import ModelPrivateAttr
-from sqlalchemy.orm.attributes import set_attribute
+from lnschema_core import File
 
 from .._docs import doc_args
 from ._core import file_to_sqm, get_session_from_kwargs
@@ -104,7 +101,6 @@ def add(  # type: ignore
     # commit metadata to database
     db_error = None
     for record in records:
-        write_objectkey(record)
         # the following ensures that queried objects (within __init__)
         # behave like queried objects, only example right now: Run
         if hasattr(record, "_ln_identity_key") and record._ln_identity_key is not None:
@@ -197,84 +193,6 @@ def prepare_error_message(records, added_records, error) -> str:
     return error_message
 
 
-def local_instance_storage_matches_local_parent(file: File):
-    storage = setup_settings.instance.storage
-
-    if file._local_filepath is not None:
-        path = file._local_filepath
-    else:
-        path = file.path()
-
-    if not isinstance(path, UPath):
-        path = path.resolve()
-
-    parents = [str(p) for p in path.parents]
-    return str(storage.root) in parents
-
-
-def get_storage_root_and_root_str(
-    root: Optional[Union[Path, UPath]] = None
-) -> Tuple[Union[Path, UPath], str]:
-    if root is None:
-        root = setup_settings.instance.storage.root
-    root_str = root.as_posix()
-    if isinstance(root, UPath):
-        root_str = root_str.rstrip("/")
-    return root, root_str
-
-
-def filepath_to_relpath(
-    root: Union[PurePath, Path], root_str: str, filepath: Union[Path, UPath]
-) -> Union[PurePath, Path]:
-    """Filepath to relative path of the root."""
-    if isinstance(root, UPath):
-        relpath = PurePath(filepath.as_posix().replace(root_str, ""))
-    else:
-        relpath = filepath.resolve().relative_to(root_str)
-
-    return relpath
-
-
-def filepath_to_objectkey(
-    record: Union[File, Folder], filepath: Union[Path, UPath]
-) -> str:
-    root, root_str = get_storage_root_and_root_str()
-
-    relpath = filepath_to_relpath(root=root, root_str=root_str, filepath=filepath)
-    # for File, _objectkey is relative path to the storage root without suffix
-    _objectkey = relpath.parent / record.name if isinstance(record, File) else relpath
-
-    return _objectkey.as_posix()
-
-
-def write_objectkey(record: sqm.SQLModel) -> None:
-    """Write to _objectkey.
-
-    An objectkey excludes the storage root and the file suffix.
-    """
-
-    def set_objectkey(record: Union[File, Folder], filepath: Union[Path, UPath]):
-        _objectkey = filepath_to_objectkey(record=record, filepath=filepath)
-        set_attribute(record, "_objectkey", _objectkey)
-
-    # _local_filepath private attribute is only added
-    # when creating File from data or Folder from folder
-    if hasattr(record, "_local_filepath"):
-        # for upsert
-        if isinstance(record._local_filepath, ModelPrivateAttr):
-            pass
-        elif record._local_filepath is None:
-            # cloud storage
-            if record._cloud_filepath is not None:
-                set_objectkey(record, record._cloud_filepath)
-            # both _cloud_filepath and _local_filepath are None fir zarr
-        # local storage
-        else:
-            # only set objectkey if it is configured
-            if local_instance_storage_matches_local_parent(record):
-                set_objectkey(record, record._local_filepath)
-
-
 def upload_data_object(file) -> None:
     """Store and add file and its linked entries."""
     file_storage_key = f"{file.id}{file.suffix}"
@@ -282,14 +200,7 @@ def upload_data_object(file) -> None:
     storage = setup_settings.instance.storage
 
     if file._local_filepath is not None:
-        # - Look for _cloud_filepath, which is only not None if the passed filepath
-        # was in the existing storage in the first place (errors within _file.py)
-        # - Look for _local_filepath and check whether it's in existing storage before
-        # trying to copy the file
-        if (file._cloud_filepath is None) and (
-            not local_instance_storage_matches_local_parent(file)
-        ):
-            store_object(file._local_filepath, file_storage_key)
+        store_object(file._local_filepath, file_storage_key)
     elif file.suffix == ".zarr" and file._memory_rep is not None:
         storagepath = storage.key_to_filepath(file_storage_key)
         print_progress = partial(print_hook, filepath=file.name)
